caxselling/aig/src/requirements.txt
--extra-index-url https://storage.openvinotoolkit.org/simple/wheels/pre-release
--extra-index-url https://storage.openvinotoolkit.org/simple/wheels/nightly
openvino_genai~=2025.3.0.0.dev #GenAI
librosa==0.10.2.post1  # Fwor Whisper
pillow==11.2.1  # Image processing for VLMs
json5==0.12.0  # For ReAct
pydantic == 2.8.2
pyYAML == 6.0.2
flask-restx == 1.3.0
flask == 2.2.5
Flask-Cors == 5.0.0
swagger-ui-bundle == 0.0.5
python_dateutil == 2.8.2
setuptools >= 75.8.0
jinja2
cython
numpy
pillow
accelerator 
chromadb # ASE for Vector DB connectivity
